+++ airflow-core/tests/unit/jobs/test_scheduler_job.py
@@ -96,6 +96,7 @@ from tests_common.test_utils.db import (
 )
 from tests_common.test_utils.mock_executor import MockExecutor
 from tests_common.test_utils.mock_operators import CustomOperator
+from tests_common.test_utils.version_compat import SQLALCHEMY_V_1_4, SQLALCHEMY_V_2_0
 from unit.listeners import dag_listener
 from unit.listeners.test_listeners import get_listener_manager
 from unit.models import TEST_DAGS_FOLDER
@@ -2089,6 +2090,105 @@ class TestSchedulerJob:
         states = [x.state for x in dr.get_task_instances(session=session)]
         assert states == ["failed", "failed"]
 
+    @conf_vars({("scheduler", "num_stuck_in_queued_retries"): "2"})
+    def test_handle_stuck_queued_tasks_reschedule_sensors(self, dag_maker, session, mock_executors):
+        """Reschedule sensors go in and out of running repeatedly using the same try_number
+        Make sure that they get three attempts per reschedule, not 3 attempts per try_number"""
+        with dag_maker("test_fail_stuck_queued_tasks_multiple_executors"):
+            EmptyOperator(task_id="op1")
+            EmptyOperator(task_id="op2", executor="default_exec")
+
+        def _queue_tasks(tis):
+            for ti in tis:
+                ti.state = "queued"
+                ti.queued_dttm = timezone.utcnow()
+            session.commit()
+
+        def _add_running_event(tis):
+            for ti in tis:
+                updated_entry = Log(
+                    dttm=timezone.utcnow(),
+                    dag_id=ti.dag_id,
+                    task_id=ti.task_id,
+                    map_index=ti.map_index,
+                    event="running",
+                    run_id=ti.run_id,
+                    try_number=ti.try_number,
+                )
+                session.add(updated_entry)
+
+        run_id = str(uuid4())
+        dr = dag_maker.create_dagrun(run_id=run_id)
+
+        tis = dr.get_task_instances(session=session)
+        _queue_tasks(tis=tis)
+        scheduler_job = Job()
+        scheduler = SchedulerJobRunner(job=scheduler_job, num_runs=0)
+        # job_runner._reschedule_stuck_task = MagicMock()
+        scheduler._task_queued_timeout = -300  # always in violation of timeout
+
+        with _loader_mock(mock_executors):
+            scheduler._handle_tasks_stuck_in_queued()
+        # If the task gets stuck in queued once, we reset it to scheduled
+        tis = dr.get_task_instances(session=session)
+        assert [x.state for x in tis] == ["scheduled", "scheduled"]
+        assert [x.queued_dttm for x in tis] == [None, None]
+
+        _queue_tasks(tis=tis)
+        log_events = [
+            x.event for x in session.scalars(select(Log).where(Log.run_id == run_id).order_by(Log.id)).all()
+        ]
+        assert log_events == [
+            "stuck in queued reschedule",
+            "stuck in queued reschedule",
+        ]
+
+        with _loader_mock(mock_executors):
+            scheduler._handle_tasks_stuck_in_queued()
+
+        log_events = [
+            x.event for x in session.scalars(select(Log).where(Log.run_id == run_id).order_by(Log.id)).all()
+        ]
+        assert log_events == [
+            "stuck in queued reschedule",
+            "stuck in queued reschedule",
+            "stuck in queued reschedule",
+            "stuck in queued reschedule",
+        ]
+        mock_executors[0].fail.assert_not_called()
+        tis = dr.get_task_instances(session=session)
+        assert [x.state for x in tis] == ["scheduled", "scheduled"]
+
+        _add_running_event(tis)  # This should "reset" the count of stuck queued
+
+        for _ in range(3):  # Should be able to be stuck 3 more times before failing
+            _queue_tasks(tis=tis)
+            with _loader_mock(mock_executors):
+                scheduler._handle_tasks_stuck_in_queued()
+            tis = dr.get_task_instances(session=session)
+
+        log_events = [
+            x.event for x in session.scalars(select(Log).where(Log.run_id == run_id).order_by(Log.id)).all()
+        ]
+        assert log_events == [
+            "stuck in queued reschedule",
+            "stuck in queued reschedule",
+            "stuck in queued reschedule",
+            "stuck in queued reschedule",
+            "running",
+            "running",
+            "stuck in queued reschedule",
+            "stuck in queued reschedule",
+            "stuck in queued reschedule",
+            "stuck in queued reschedule",
+            "stuck in queued tries exceeded",
+            "stuck in queued tries exceeded",
+        ]
+
+        mock_executors[0].fail.assert_not_called()  # just demoing that we don't fail with executor method
+        states = [x.state for x in dr.get_task_instances(session=session)]
+        assert states == ["failed", "failed"]
+
     def test_revoke_task_not_imp_tolerated(self, dag_maker, session, caplog):
         """Test that if executor no implement revoke_task then we don't blow up."""
         with dag_maker("test_fail_stuck_queued_tasks"):
@@ -2275,21 +2375,21 @@ class TestSchedulerJob:
         assert dr.run_id is not None
         assert dr.state == State.RUNNING
         assert dr.span_status == SpanStatus.ACTIVE
-        assert self.job_runner.active_spans.get(dr.run_id) is None
+        assert self.job_runner.active_spans.get("dr:" + str(dr.id)) is None
 
-        assert self.job_runner.active_spans.get(ti.key) is None
+        assert self.job_runner.active_spans.get("ti:" + ti.id) is None
         assert ti.state == ti_state
         assert ti.span_status == SpanStatus.ACTIVE
 
         self.job_runner._recreate_unhealthy_scheduler_spans_if_needed(dr, session)
 
-        assert self.job_runner.active_spans.get(dr.run_id) is not None
+        assert self.job_runner.active_spans.get("dr:" + str(dr.id)) is not None
 
         if final_ti_span_status == SpanStatus.ACTIVE:
-            assert self.job_runner.active_spans.get(ti.key) is not None
+            assert self.job_runner.active_spans.get("ti:" + ti.id) is not None
             assert len(self.job_runner.active_spans.get_all()) == 2
         else:
-            assert self.job_runner.active_spans.get(ti.key) is None
+            assert self.job_runner.active_spans.get("ti:" + ti.id) is None
             assert len(self.job_runner.active_spans.get_all()) == 1
 
         assert dr.span_status == SpanStatus.ACTIVE
@@ -2329,22 +2429,22 @@ class TestSchedulerJob:
         dr_span = Trace.start_root_span(span_name="dag_run_span", start_as_current=False)
         ti_span = Trace.start_child_span(span_name="ti_span", start_as_current=False)
 
-        self.job_runner.active_spans.set(dr.run_id, dr_span)
-        self.job_runner.active_spans.set(ti.key, ti_span)
+        self.job_runner.active_spans.set("dr:" + str(dr.id), dr_span)
+        self.job_runner.active_spans.set("ti:" + ti.id, ti_span)
 
         assert dr.span_status == SpanStatus.SHOULD_END
         assert ti.span_status == SpanStatus.SHOULD_END
 
-        assert self.job_runner.active_spans.get(dr.run_id) is not None
-        assert self.job_runner.active_spans.get(ti.key) is not None
+        assert self.job_runner.active_spans.get("dr:" + str(dr.id)) is not None
+        assert self.job_runner.active_spans.get("ti:" + ti.id) is not None
 
         self.job_runner._end_spans_of_externally_ended_ops(session)
 
         assert dr.span_status == SpanStatus.ENDED
         assert ti.span_status == SpanStatus.ENDED
 
-        assert self.job_runner.active_spans.get(dr.run_id) is None
-        assert self.job_runner.active_spans.get(ti.key) is None
+        assert self.job_runner.active_spans.get("dr:" + str(dr.id)) is None
+        assert self.job_runner.active_spans.get("ti:" + ti.id) is None
 
     @pytest.mark.parametrize(
         "state, final_span_status",
@@ -2386,14 +2486,14 @@ class TestSchedulerJob:
         dr_span = Trace.start_root_span(span_name="dag_run_span", start_as_current=False)
         ti_span = Trace.start_child_span(span_name="ti_span", start_as_current=False)
 
-        self.job_runner.active_spans.set(dr.run_id, dr_span)
-        self.job_runner.active_spans.set(ti.key, ti_span)
+        self.job_runner.active_spans.set("dr:" + str(dr.id), dr_span)
+        self.job_runner.active_spans.set("ti:" + ti.id, ti_span)
 
         assert dr.span_status == SpanStatus.ACTIVE
         assert ti.span_status == SpanStatus.ACTIVE
 
-        assert self.job_runner.active_spans.get(dr.run_id) is not None
-        assert self.job_runner.active_spans.get(ti.key) is not None
+        assert self.job_runner.active_spans.get("dr:" + str(dr.id)) is not None
+        assert self.job_runner.active_spans.get("ti:" + ti.id) is not None
         assert len(self.job_runner.active_spans.get_all()) == 2
 
         self.job_runner._end_active_spans(session)
@@ -2401,8 +2501,8 @@ class TestSchedulerJob:
         assert dr.span_status == final_span_status
         assert ti.span_status == final_span_status
 
-        assert self.job_runner.active_spans.get(dr.run_id) is None
-        assert self.job_runner.active_spans.get(ti.key) is None
+        assert self.job_runner.active_spans.get("dr:" + str(dr.id)) is None
+        assert self.job_runner.active_spans.get("ti:" + ti.id) is None
         assert len(self.job_runner.active_spans.get_all()) == 0
 
     def test_dagrun_timeout_verify_max_active_runs(self, dag_maker):
@@ -3354,7 +3454,7 @@ class TestSchedulerJob:
 
         # Now let's say the DAG got updated (new task got added)
         BashOperator(task_id="bash_task_1", dag=dag, bash_command="echo hi")
-        SerializedDagModel.write_dag(dag=dag, bundle_name="testing")
+        SerializedDagModel.write_dag(dag=dag, bundle_name="testing", session=session)
 
         dag_version_2 = DagVersion.get_latest_version(dr.dag_id, session=session)
         assert dag_version_2 != dag_version_1
@@ -3368,15 +3468,24 @@ class TestSchedulerJob:
         assert dr.dag_versions[-1].id == dag_version_2.id
         assert len(self.job_runner.scheduler_dag_bag.get_dag(dr, session).tasks) == 2
 
-        tis_count = (
-            session.query(func.count(TaskInstance.task_id))
-            .filter(
-                TaskInstance.dag_id == dr.dag_id,
-                TaskInstance.logical_date == dr.logical_date,
-                TaskInstance.state == State.SCHEDULED,
+        if SQLALCHEMY_V_1_4:
+            tis_count = (
+                session.query(func.count(TaskInstance.task_id))
+                .filter(
+                    TaskInstance.dag_id == dr.dag_id,
+                    TaskInstance.logical_date == dr.logical_date,
+                    TaskInstance.state == State.SCHEDULED,
+                )
+                .scalar()
             )
-            .scalar()
-        )
+        if SQLALCHEMY_V_2_0:
+            tis_count = session.scalar(
+                select(func.count(TaskInstance.task_id)).where(
+                    TaskInstance.dag_id == dr.dag_id,
+                    TaskInstance.logical_date == dr.logical_date,
+                    TaskInstance.state == State.SCHEDULED,
+                )
+            )
         assert tis_count == 2
 
         latest_dag_version = DagVersion.get_latest_version(dr.dag_id, session=session)
@@ -4715,6 +4824,7 @@ class TestSchedulerJob:
             to_date=to_date,
             max_active_runs=3,
             reverse=False,
+            triggering_user_name="test_user",
             dag_run_conf={},
         )
         dag1_running_count = (
@@ -4819,6 +4929,7 @@ class TestSchedulerJob:
             to_date=to_date,
             max_active_runs=3,
             reverse=False,
+            triggering_user_name="test_user",
             dag_run_conf={},
         )
 
@@ -4894,6 +5005,7 @@ class TestSchedulerJob:
             to_date=to_date,
             max_active_runs=3,
             reverse=False,
+            triggering_user_name="test_user",
             dag_run_conf={},
         )
         dag1_non_b_running, dag1_b_running, total_running = _running_counts()
@@ -5009,6 +5121,7 @@ class TestSchedulerJob:
             to_date=to_date,
             max_active_runs=3,
             reverse=False,
+            triggering_user_name="test_user",
             dag_run_conf={},
         )
         dag1_non_b_running, dag1_b_running, total_running = _running_counts()
@@ -5115,6 +5228,7 @@ class TestSchedulerJob:
             to_date=to_date,
             max_active_runs=3,
             reverse=False,
+            triggering_user_name="test_user",
             dag_run_conf={},
         )
         dag1_non_b_running, dag1_b_running, total_running = _running_counts()
@@ -5263,6 +5377,7 @@ class TestSchedulerJob:
             to_date=to_date,
             max_active_runs=3,
             reverse=False,
+            triggering_user_name="test_user",
             dag_run_conf={},
         )
         dag1_non_b_running, dag1_b_running, total_running = _running_counts()
@@ -6642,6 +6757,7 @@ class TestSchedulerJobQueriesCount:
 
                     self.job_runner._run_scheduler_loop()
 
+    @pytest.mark.flaky(reruns=3, reruns_delay=5)
     @pytest.mark.parametrize(
         "expected_query_counts, dag_count, task_count, start_ago, schedule, shape",
         [
@@ -6735,6 +6851,7 @@ def test_mark_backfills_completed(dag_maker, session):
         to_date=pendulum.parse("2021-01-03"),
         max_active_runs=10,
         reverse=False,
+        triggering_user_name="test_user",
         dag_run_conf={},
     )
     session.expunge_all()