+++ helm-tests/tests/helm_tests/security/test_rbac.py
@@ -47,7 +47,6 @@ DEPLOYMENT_NO_RBAC_NO_SA_KIND_NAME_TUPLES = [
     ("Secret", "test-rbac-broker-url"),
     ("Secret", "test-rbac-fernet-key"),
     ("Secret", "test-rbac-redis-password"),
-    ("Secret", "test-rbac-webserver-secret-key"),
     ("Job", "test-rbac-create-user"),
     ("Job", "test-rbac-run-airflow-migrations"),
     ("CronJob", "test-rbac-cleanup"),
@@ -122,6 +121,7 @@ class TestRBAC:
                     ("Service", "test-rbac-api-server"),
                     ("Deployment", "test-rbac-api-server"),
                     ("Deployment", "test-rbac-dag-processor"),
+                    ("Secret", "test-rbac-api-secret-key"),
                     ("Secret", "test-rbac-jwt-secret"),
                 )
             )
@@ -133,6 +133,7 @@ class TestRBAC:
                 (
                     ("Service", "test-rbac-webserver"),
                     ("Deployment", "test-rbac-webserver"),
+                    ("Secret", "test-rbac-webserver-secret-key"),
                 )
             )
             if sa: