+++ providers/apache/flink/src/airflow/providers/apache/flink/sensors/flink_kubernetes.py
@@ -23,16 +23,11 @@ from typing import TYPE_CHECKING
 from kubernetes import client
 
 from airflow.exceptions import AirflowException
-from airflow.providers.apache.flink.version_compat import AIRFLOW_V_3_0_PLUS
+from airflow.providers.apache.flink.version_compat import BaseSensorOperator
 from airflow.providers.cncf.kubernetes.hooks.kubernetes import KubernetesHook
 
-if AIRFLOW_V_3_0_PLUS:
-    from airflow.sdk import BaseSensorOperator
-else:
-    from airflow.sensors.base import BaseSensorOperator  # type: ignore[no-redef]
-
 if TYPE_CHECKING:
-    from airflow.utils.context import Context
+    from airflow.providers.apache.flink.version_compat import Context
 
 
 class FlinkKubernetesSensor(BaseSensorOperator):